+++ python/cpp/pytext.cpp
@@ -69,9 +69,51 @@ NVStrings* strings_from_list(PyObject* listObj)
 
 //
 //
+static PyObject* n_tokenize( PyObject* self, PyObject* args )
+{
+    PyObject* pystrs = PyTuple_GetItem(args,0);
+    NVStrings* strs = strings_from_object(pystrs);
+    if( strs==0 )
+        Py_RETURN_NONE;
+
+    const char* delimiter = " ";
+    PyObject* argDelim = PyTuple_GetItem(args,1);
+    if( argDelim != Py_None )
+        delimiter = PyUnicode_AsUTF8(argDelim);
+
+    Py_BEGIN_ALLOW_THREADS
+    strs = NVText::tokenize(*strs,delimiter);
+    Py_END_ALLOW_THREADS
+    if( strs==0 )
+        Py_RETURN_NONE;
+    return PyLong_FromVoidPtr((void*)strs);
+}
+
+//
+static PyObject* n_tokenize_multi( PyObject* self, PyObject* args )
+{
+    PyObject* pystrs = PyTuple_GetItem(args,0);
+    NVStrings* strs = strings_from_object(pystrs);
+    if( strs==0 )
+        Py_RETURN_NONE;
+
+    PyObject* pydelims = PyTuple_GetItem(args,1);
+    NVStrings* delims = strings_from_object(pydelims);
+    if( delims==0 )
+        Py_RETURN_NONE;
+
+    Py_BEGIN_ALLOW_THREADS
+    strs = NVText::tokenize(*strs,*delims);
+    Py_END_ALLOW_THREADS
+    if( strs==0 )
+        Py_RETURN_NONE;
+    return PyLong_FromVoidPtr((void*)strs);
+}
+
+//
 static PyObject* n_unique_tokens( PyObject* self, PyObject* args )
 {
-    PyObject* pystrs = PyTuple_GetItem(args,0); // only one parm expected
+    PyObject* pystrs = PyTuple_GetItem(args,0);
     NVStrings* strs = strings_from_object(pystrs);
     if( strs==0 )
         Py_RETURN_NONE;
@@ -493,6 +535,8 @@ static PyObject* n_edit_distance( PyObject* self, PyObject* args )
 
 //
 static PyMethodDef s_Methods[] = {
+    { "n_tokenize", n_tokenize, METH_VARARGS, "" },
+    { "n_tokenize_multi", n_tokenize_multi, METH_VARARGS, "" },
     { "n_unique_tokens", n_unique_tokens, METH_VARARGS, "" },
     { "n_token_count", n_token_count, METH_VARARGS, "" },
     { "n_contains_strings", n_contains_strings, METH_VARARGS, "" },