+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version jump from 0.3->0.7 is to align with other RAPIDS projects.
 
 ## New Features
 
+- PR #295 Added nvtext.tokenize
 - PR #281 Added get_info method to show data about strings
 - PR #273 Added tokens_counts method to nvtext
 - PR #265 Added is_empty method
@@ -18,7 +19,9 @@ Version jump from 0.3->0.7 is to align with other RAPIDS projects.
 - PR #172 Added to/from boolean conversion methods
 - PR #171 Added conversion to/from subset of ISO8601 format
 - PR #230 Add ngrams function to nvtext module
+- PR #285 Add build script for nightly docs
 - PR #286 Add local build script to cuStrings
+
 
 ## Improvements